pyproject.toml
[tool.poetry]
name = "parquet-s3-meta"
version = "0.1.0"
description = ""
authors = ["Aditya Rachman Putra <adityarputra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.1"
awscli = "^1.32.57"
faker = "^24.0.0"
polars = "^0.20.14"
boto3 = "^1.34.57"
python-dotenv = "^1.0.1"
pyarrow = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
